+++ python/cugraph/jaccard/test_jaccard.py
@@ -58,7 +58,7 @@ def cugraph_call(M, edgevals=False):
 
     # cugraph Jaccard Call
     t1 = time.time()
-    df = cugraph.nvJaccard(G)
+    df = cugraph.jaccard(G)
     t2 = time.time() - t1
     print('Time : '+str(t2))
 
@@ -145,3 +145,52 @@ def test_jaccard_edgevals(graph_file):
 
     print("Mismatches:  %d" % err)
     assert err == 0
+
+
+@pytest.mark.parametrize('graph_file', DATASETS)
+def test_jaccard_two_hop(graph_file):
+    M = read_mtx_file(graph_file)
+    M = M.tocsr()
+    Gnx = nx.DiGraph(M).to_undirected()
+    G = cugraph.Graph()
+    row_offsets = cudf.Series(M.indptr)
+    col_indices = cudf.Series(M.indices)
+    G.add_adj_list(row_offsets, col_indices, None)
+    pairs = G.get_two_hop_neighbors()
+    nx_pairs = []
+    for i in range(len(pairs)):
+        nx_pairs.append((pairs['first'][i], pairs['second'][i]))
+    preds = nx.jaccard_coefficient(Gnx, nx_pairs)
+    nx_coeff = []
+    for u, v, p in preds:
+        nx_coeff.append(p)
+    df = cugraph.jaccard(G, pairs['first'], pairs['second'])
+    assert len(nx_coeff) == len(df)
+    for i in range(len(df)):
+        diff = abs(nx_coeff[i] - df['jaccard_coeff'][i])
+        assert diff < 1.0e-6
+
+
+@pytest.mark.parametrize('graph_file', DATASETS)
+def test_jaccard_two_hop_edge_vals(graph_file):
+    M = read_mtx_file(graph_file)
+    M = M.tocsr()
+    Gnx = nx.DiGraph(M).to_undirected()
+    G = cugraph.Graph()
+    row_offsets = cudf.Series(M.indptr)
+    col_indices = cudf.Series(M.indices)
+    values = cudf.Series(M.data)
+    G.add_adj_list(row_offsets, col_indices, values)
+    pairs = G.get_two_hop_neighbors()
+    nx_pairs = []
+    for i in range(len(pairs)):
+        nx_pairs.append((pairs['first'][i], pairs['second'][i]))
+    preds = nx.jaccard_coefficient(Gnx, nx_pairs)
+    nx_coeff = []
+    for u, v, p in preds:
+        nx_coeff.append(p)
+    df = cugraph.jaccard(G, pairs['first'], pairs['second'])
+    assert len(nx_coeff) == len(df)
+    for i in range(len(df)):
+        diff = abs(nx_coeff[i] - df['jaccard_coeff'][i])
+        assert diff < 1.0e-6